comment works

+++ src/app/components/forum/forum.component.ts
@@ -1,21 +1,9 @@
 import { Component, Injectable, OnInit } from '@angular/core';
-import { PostServiceService } from 'src/app/Service/post-service.service';
 import { Router } from '@angular/router';
-
-
-export class Posts{
-
-  constructor(
-    public postId : number,
-    public titlePost : string,
-    public descriptionPost : string,
-    public nbComment : number,
-    public datePost : Date,
-    public imagePost : string
-  ){
-
-  }
-}
+import { DictionaryService } from 'src/app/Service/dictionaryService/dictionary.service';
+import { Posts } from 'src/app/models/Posts';
+import { Dictionary } from 'src/app/models/Dictionary';
+import { PostServiceService } from 'src/app/Service/PostService/post-service.service';
 
 
 @Component({
@@ -28,10 +16,12 @@ export class Posts{
 export class ForumComponent implements OnInit {
 
    listPosts: Posts[] = [];
-
+   bestPosts : Posts[]=[];
+   listWords : Dictionary[] = [];
 
   constructor(
     public postService : PostServiceService,
+    public dictionaryService : DictionaryService,
     public router : Router
     
     ) { }
@@ -43,13 +33,21 @@ export class ForumComponent implements OnInit {
         this.listPosts=response;
       }
     );
+
+    this.dictionaryService.retrieveAllWords().subscribe(
+      response=>{
+        console.log(response);
+        this.listWords=response;
+      }
+    )
+   
   }
 
 
    message ="";
    deletePost(id:number){
-
-    console.log(`delete post ${id}`);
+if(window.confirm("Are you sure you want to delete this post !!")){
+  console.log(`delete post ${id}`);
     this.postService.deletePost(id).subscribe(
         response=>{
           console.log(response);
@@ -64,6 +62,50 @@ export class ForumComponent implements OnInit {
        }
         
   )
+}
+    
    
    }
+
+
+   deleteWord(id:number){
+    if(window.confirm("Are you sure you want to delete this Word !!")){
+      console.log(`delete word ${id}`);
+        this.dictionaryService.deleteWord(id).subscribe(
+            response=>{
+              console.log(response);
+              this.message = `  word ${id} deleted successfully `;
+             this.dictionaryService.retrieveAllWords().subscribe(
+                response => {
+                  console.log(response);
+                this.listWords=response;
+               }
+             );
+              
+           }
+            
+      )
+    }
+        
+       
+       }
+
+   updatePost(id:number){
+    console.log(`update post ${id}`)
+    this.router.navigate(['post',id])
+  }
+
+  addPost(){
+    
+    this.router.navigate(['addpost'])
+  }
+
+  addWord(){
+    this.router.navigate(["addictionary"])
+  }
+
+  updateWord(id:number){
+   console.log(`update word ${id}`)
+   this.router.navigate(['updatedictionary',id])
+  }
 }
